pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "deckr-controller"
version = "0.1.0"
description = "Deckr controller runtime with config, routing, rendering, and device management"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Kaj Siebert", email = "kaj@k-si.com" }
]
requires-python = ">=3.11,<4.0.0"
keywords = ["deckr", "stream-deck", "controller", "automation", "plugins"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Hardware",
]
dependencies = [
    "aiomqtt>=2.0.0,<3.0.0",
    "cairosvg>=2.8.2,<3.0.0",
    "deckr>=0.1.0,<0.2.0",
    "deckr-fonts>=0.2.0",
    "httpx>=0.28.1,<0.29.0",
    "invariant-core==0.3.0",
    "invariant-gfx==0.2.0",
    "justmyresource[icons]>=1.1.1,<2.0.0",
    "justmyresource-lucide>=0.575.0",
    "justmytype>=0.3.0,<0.4.0",
    "pillow>=12.1.0,<13.0.0",
    "platformdirs>=4.5.1,<5.0.0",
    "python-decouple>=3.8,<4.0",
    "pyyaml>=6.0.3,<7.0.0",
    "simpleeval>=1.0.3,<2.0.0",
    "tinydb>=4.8.2,<5.0.0",
    "watchfiles>=1.1.1,<2.0.0",
    "websockets>=12.0,<16.0",
]

[project.scripts]
deckr = "deckr.controller:main"
deckr-device-manager = "deckr.controller:device_manager_main"

[project.urls]
Documentation = "https://github.com/kws/deckr-controller#readme"
Homepage = "https://github.com/kws/deckr-controller"
Repository = "https://github.com/kws/deckr-controller"
"Bug Tracker" = "https://github.com/kws/deckr-controller/issues"
"Changelog" = "https://github.com/kws/deckr-controller/releases"

[dependency-groups]
dev = [
    "pytest>=9.0.2,<10.0.0",
    "pytest-asyncio>=1.3.0,<2.0.0",
    "pytest-cov>=6.0.0,<7.0.0",
    "ruff>=0.14.14,<0.15.0",
]

[tool.uv.sources]
deckr = { path = "../deckr", editable = true }

[tool.hatch.build.targets.wheel]
packages = ["src/deckr"]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = [
    "B",
    "E",
    "F",
    "I",
    "SIM",
    "UP",
    "W",
]
ignore = [
    "E203",
    "E501",
    "SIM102",
    "SIM105",
]

[tool.pytest.ini_options]
addopts = [
    "--cov=src/deckr",
    "--cov-report=term-missing",
]
testpaths = ["tests"]
